Move normalizeStringList into plugin option utils
