+++ src/plugin/src/options/line-comment-options.js
@@ -1,6 +1,8 @@
 import { getCachedValue } from "../options/options-cache.js";
-import { coercePositiveIntegerOption } from "../utils/option-utils.js";
-import { normalizeStringList } from "../../../shared/string-utils.js";
+import {
+    coercePositiveIntegerOption,
+    normalizeStringList
+} from "../utils/option-utils.js";
 
 const DEFAULT_LINE_COMMENT_BANNER_MIN_SLASHES = 5;
 const DEFAULT_LINE_COMMENT_BANNER_AUTOFILL_THRESHOLD = 4;
